requirements.txt
django
twitter
PyYAML
BeautifulSoup
